+++ bin/run.rb
@@ -1,4 +1,4 @@
 require_relative '../config/environment'
 
-binding.pry
+#binding.pry
 puts "HELLO WORLD"
