+++ db/migrate/20190327131936_add_job_saved_column.rb
@@ -1,0 +1,5 @@
+class AddJobSavedColumn < ActiveRecord::Migration[5.0]
+  def change
+  	add_column :jobs, :job_saved, :boolean
+  end
+end
