+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20190326153919) do
+ActiveRecord::Schema.define(version: 20190327131936) do
 
   create_table "jobs", force: :cascade do |t|
     t.string   "position_hours"
@@ -24,6 +24,7 @@ ActiveRecord::Schema.define(version: 20190326153919) do
     t.string   "how_to_apply"
     t.datetime "updated_at",     null: false
     t.string   "job_uniq_id"
+    t.boolean  "job_saved"
   end
 
   create_table "user_jobs", force: :cascade do |t|
